feat: expose lightweight API for resolving a command path
